# src/api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "LumiRun API",
    "version" : "1.0.0",
    "title" : "LumiRun",
    "contact" : {
      "email" : "jps@Lumi.education"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "H5P"
  }, {
    "name" : "Analytics"
  }, {
    "name" : "Static"
  } ],
  "paths" : {
    "/v0/h5p/render" : {
      "get" : {
        "tags" : [ "H5P" ],
        "summary" : "Renders H5P",
        "operationId" : "runh5p",
        "description" : "Renders H5P\n",
        "produces" : [ "text/html" ],
        "parameters" : [ {
          "in" : "query",
          "name" : "id",
          "description" : "H5P id",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "Rendered H5P Page"
          },
          "404" : {
            "description" : "H5P not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/v0/h5p" : {
      "get" : {
        "tags" : [ "H5P" ],
        "summary" : "returns H5P parameters and metadata",
        "operationId" : "geth5p",
        "description" : "returns H5P parameters and metadata\n",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "query",
          "name" : "id",
          "description" : "H5P id",
          "required" : true,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "H5P metadata and parameters",
            "schema" : {
              "$ref" : "#/definitions/H5P"
            }
          },
          "404" : {
            "description" : "H5P not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "H5P" ],
        "summary" : "Upload a H5P",
        "operationId" : "addh5p",
        "description" : "Upload a valid H5P to Lumi Run",
        "consumes" : [ "multipart/form-data" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "formData",
          "name" : "h5p",
          "type" : "file",
          "description" : "The file to upload."
        } ],
        "responses" : {
          "201" : {
            "description" : "H5P added",
            "schema" : {
              "type" : "object",
              "properties" : {
                "analytics_id" : {
                  "type" : "string",
                  "format" : "uuid",
                  "example" : "aY64Fg"
                },
                "h5p_id" : {
                  "type" : "integer",
                  "format" : "uuid",
                  "example" : 123456
                }
              }
            }
          },
          "400" : {
            "description" : "invalid H5P",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/v0/analytics" : {
      "get" : {
        "tags" : [ "Analytics" ],
        "summary" : "Returns analytics for a H5P",
        "operationId" : "getAnalytics",
        "description" : "Returns analytics\n",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "query",
          "name" : "id",
          "description" : "Analytics id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Rendered H5P Page",
            "schema" : {
              "type" : "object",
              "properties" : {
                "_id" : {
                  "type" : "string",
                  "format" : "uuid",
                  "example" : "aY64Fg"
                },
                "accessed_at" : {
                  "type" : "number",
                  "format" : "Date",
                  "example" : 134459459
                },
                "created_at" : {
                  "type" : "number",
                  "format" : "Date",
                  "example" : 132445645
                },
                "data" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Analytics"
                  }
                },
                "h5p_id" : {
                  "type" : "integer",
                  "format" : "uuid",
                  "example" : 125486
                }
              }
            }
          },
          "404" : {
            "description" : "Analytics not found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Analytics" ],
        "summary" : "Saves data to analytics",
        "operationId" : "postAnalytics",
        "description" : "Returns analytics\n",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "query",
          "name" : "h5p_id",
          "description" : "H5P id",
          "required" : true,
          "type" : "integer"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Analytics"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Analytics processed"
          },
          "404" : {
            "description" : "H5P does not exist",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Analytics" : {
      "type" : "object",
      "required" : [ "contentId", "score", "maxScore", "opened", "finished" ],
      "properties" : {
        "contentId" : {
          "type" : "string",
          "example" : "klmdfg"
        },
        "score" : {
          "type" : "string",
          "example" : 0
        },
        "maxScore" : {
          "type" : "string",
          "example" : 1
        },
        "opened" : {
          "type" : "string",
          "example" : 1577049015
        },
        "finished" : {
          "type" : "string",
          "example" : 1577049017
        }
      }
    },
    "H5P" : {
      "type" : "object",
      "required" : [ "id", "metadata", "parameters" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "uuid",
          "example" : 123456
        },
        "metadata" : {
          "type" : "object",
          "example" : "TBD"
        },
        "parameters" : {
          "type" : "object",
          "example" : "TBD"
        }
      }
    },
    "Error" : {
      "type" : "object",
      "required" : [ "code", "status", "message" ],
      "properties" : {
        "code" : {
          "type" : "string",
          "example" : "h5p-not-found"
        },
        "status" : {
          "type" : "integer",
          "example" : 404
        },
        "message" : {
          "type" : "string",
          "example" : "H5P was not found"
        }
      }
    }
  }
}